+++ aw-arch/src/main/java/com/answufeng/arch/base/BaseActivity.kt
@@ -4,6 +4,13 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.viewbinding.ViewBinding
 
+/**
+ * 基础 Activity 基类
+ *
+ * 提供了 ViewBinding 支持，简化了 ViewBinding 的初始化和生命周期管理
+ *
+ * @param VB ViewBinding 类型
+ */
 abstract class BaseActivity<VB : ViewBinding> : AppCompatActivity() {
 
     private var _binding: VB? = null