+++ aw-arch/src/main/java/com/answufeng/arch/hilt/HiltMvvmBottomSheetDialogFragment.kt
@@ -8,11 +8,22 @@ import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import androidx.viewbinding.ViewBinding
 import com.answufeng.arch.base.MvvmViewModel
-import com.answufeng.arch.base.MvvmViewModel.UIEvent
+import com.answufeng.arch.base.MvvmViewModel.UiEvent
 import com.answufeng.arch.mvvm.MvvmView
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import kotlinx.coroutines.launch
 
+/**
+ * Hilt 支持的 MVVM 架构 BottomSheetDialogFragment 基类
+ *
+ * 适用于使用 Hilt 进行依赖注入的 MVVM 模式 BottomSheetDialogFragment，提供了 ViewBinding 支持和 UI 事件处理
+ *
+ * @param VB ViewBinding 类型
+ * @param VM ViewModel 类型，必须继承 [MvvmViewModel]
+ *
+ * @see MvvmViewModel
+ * @see MvvmView
+ */
 abstract class HiltMvvmBottomSheetDialogFragment<VB : ViewBinding, VM : MvvmViewModel> : BottomSheetDialogFragment(), MvvmView {
 
     private var _binding: VB? = null
@@ -40,7 +51,7 @@ abstract class HiltMvvmBottomSheetDialogFragment<VB : ViewBinding, VM : MvvmView
     open fun initObservers() {
         lifecycleScope.launch {
             repeatOnLifecycle(androidx.lifecycle.Lifecycle.State.STARTED) {
-                viewModel.uiEvent.collect { onUIEvent(it) }
+                viewModel.UiEvent.collect { onUiEvent(it) }
             }
         }
     }