+++ aw-arch/src/main/java/com/answufeng/arch/mvvm/MvvmBottomSheetDialogFragment.kt
@@ -9,7 +9,7 @@ import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import androidx.viewbinding.ViewBinding
 import com.answufeng.arch.base.MvvmViewModel
-import com.answufeng.arch.base.MvvmViewModel.UIEvent
+import com.answufeng.arch.base.MvvmViewModel.UiEvent
 import com.answufeng.arch.ext.inferViewModelClass
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import kotlinx.coroutines.launch
@@ -42,7 +42,7 @@ abstract class MvvmBottomSheetDialogFragment<VB : ViewBinding, VM : MvvmViewMode
     open fun initObservers() {
         lifecycleScope.launch {
             repeatOnLifecycle(androidx.lifecycle.Lifecycle.State.STARTED) {
-                viewModel.uiEvent.collect { onUIEvent(it) }
+                viewModel.UiEvent.collect { onUiEvent(it) }
             }
         }
     }