+++ aw-arch/src/main/java/com/answufeng/arch/mvvm/MvvmFragment.kt
@@ -11,7 +11,7 @@ import androidx.lifecycle.repeatOnLifecycle
 import androidx.viewbinding.ViewBinding
 import com.answufeng.arch.base.LazyLoadHelper
 import com.answufeng.arch.base.MvvmViewModel
-import com.answufeng.arch.base.MvvmViewModel.UIEvent
+import com.answufeng.arch.base.MvvmViewModel.UiEvent
 import com.answufeng.arch.ext.inferViewModelClass
 import kotlinx.coroutines.launch
 
@@ -66,7 +66,7 @@ abstract class MvvmFragment<VB : ViewBinding, VM : MvvmViewModel> : Fragment(), 
     open fun initObservers() {
         lifecycleScope.launch {
             repeatOnLifecycle(androidx.lifecycle.Lifecycle.State.STARTED) {
-                viewModel.uiEvent.collect { onUIEvent(it) }
+                viewModel.UiEvent.collect { onUiEvent(it) }
             }
         }
     }